// source/frontend/src/models/api/generated/ApiGen_Concepts_Role.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Base_BaseAudit } from './ApiGen_Base_BaseAudit';
import { ApiGen_Concepts_RoleClaim } from './ApiGen_Concepts_RoleClaim';

// LINK: @backend/apimodels/Models/Concepts/Role/RoleModel.cs
export interface ApiGen_Concepts_Role extends ApiGen_Base_BaseAudit {
  id: number;
  roleUid: string | null;
  keycloakGroupId: string | null;
  name: string | null;
  description: string | null;
  isPublic: boolean;
  isDisabled: boolean;
  displayOrder: number;
  roleClaims: ApiGen_Concepts_RoleClaim[] | null;
}